Split HLC coordinator supporting types into separate files
